% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{as_abbrev}
\alias{as_abbrev}
\title{Abbreviate country and regional names to ISO-3X.}
\usage{
as_abbrev(x, georgia = "country", abbrev = NA)
}
\arguments{
\item{x}{(character vector) The full names of countries and subcountry units.}

\item{georgia}{(character scalar) Whether "georgia" is a country or a US state ("state"). Defaults to "country".}

\item{abbrev}{(data.framea) Optional: A data.frame of abbreviations to use.}
}
\description{
To enable easier merging of datasets of international data. You need to download the countrylist.csv file yourself.
}
\examples{
as_abbrev(c("Denmark", "USA", "India", "Thailand", "Russia", "Korea, South"))
}

