% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{as_long}
\alias{as_long}
\title{Get full country names from ISO-3X.}
\usage{
as_long(x, abbrev = NA)
}
\arguments{
\item{x}{(character vector) The ISO-3X codes.}
}
\description{
To enable easier merging of datasets of international data. You need to download the countrylist.csv file yourself.
}
\examples{
as_long(c("DNK", "USA", "IND", "THA", "RUS", "KOR"))
}

