% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_abbrev2}
\alias{get_abbrev2}
\title{Get the abbreviations data.frame.}
\usage{
get_abbrev2()
}
\arguments{
\item{x}{(character vector) The ISO-3X codes.}
}
\description{
Fetches the abbreviation dictionary and returns it to the user. Useful if one wants to alter it and pass the modified version to as_abbrev() or as_long().
}
\examples{
get_abbrev()
}

